Adding a TodoTaskManager for managing the TodoTasks. Initially empty. Setting up unit tests for testing the TodoTaskManager. Adding "initial empty" test case.
