+++ SBGTodoListMVC/SBGTodoListMVC/Model/TodoTaskManager.swift
@@ -1,0 +1,44 @@
+//
+//  TodoTaskManager.swift
+//  SBGTodoListMVC
+//
+//  Created by Tim Fuqua on 8/25/18.
+//  Copyright © 2018 SBGCodeTest. All rights reserved.
+//
+
+import Foundation
+import CoreData
+
+
+// MARK:- TodoTaskManager
+class TodoTaskManager {
+    
+    private var allTasks: [NSManagedObject] = []
+    
+}
+
+
+// MARK:- initialize
+extension TodoTaskManager {
+    
+}
+
+
+// MARK:- get
+extension TodoTaskManager {
+    func getAllTasks() -> [NSManagedObject] {
+        return allTasks
+    }
+}
+
+
+// MARK:- add
+extension TodoTaskManager {
+    
+}
+
+
+// MARK:- remove
+extension TodoTaskManager {
+    
+}
